rag-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ollama
sentence-transformers
psycopg2-binary
